requirements.txt
altgraph==0.17
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
future==0.18.2
idna==3.1
pefile==2019.4.18
Pillow==8.1.2
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
pywin32-ctypes==0.2.0
requests==2.25.1
urllib3==1.26.4
